ui: Move useful parts to `httpserver`
In the future, we want to remove the ui build artifacts from the
repository. This will break the usages of

```sh
go get github.com/prometheus/alertmanager/ui
```

That is exactly as expected, since the ui can't function correctly
without the ui artefacts.

This change attempts to make the transition a bit less painful.
