+++ ui/web.go
@@ -15,15 +15,9 @@ package ui
 
 import (
 	"embed"
-	"fmt"
 	"io/fs"
-	"log/slog"
 	"net/http"
-	_ "net/http/pprof" // Comment this line to disable pprof endpoint.
-	"path"
-	"strings"
 
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/prometheus/common/route"
 )
 
@@ -31,9 +25,7 @@ import (
 var asset embed.FS
 
 // Register registers handlers to serve files for the web interface.
-func Register(r *route.Router, reloadCh chan<- chan error, logger *slog.Logger) {
-	r.Get("/metrics", promhttp.Handler().ServeHTTP)
-
+func Register(r *route.Router) {
 	appFS, err := fs.Sub(asset, "app")
 	if err != nil {
 		panic(err) // During build step, we did not embed a directory named `app`.
@@ -58,43 +50,6 @@ func Register(r *route.Router, reloadCh chan<- chan error, logger *slog.Logger) 
 		disableCaching(w)
 		fs.ServeHTTP(w, req)
 	})
-
-	r.Post("/-/reload", func(w http.ResponseWriter, req *http.Request) {
-		errc := make(chan error)
-		defer close(errc)
-
-		reloadCh <- errc
-		if err := <-errc; err != nil {
-			http.Error(w, fmt.Sprintf("failed to reload config: %s", err), http.StatusInternalServerError)
-		}
-	})
-
-	r.Get("/-/healthy", func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusOK)
-		fmt.Fprintf(w, "OK")
-	})
-	r.Head("/-/healthy", func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusOK)
-	})
-	r.Get("/-/ready", func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusOK)
-		fmt.Fprintf(w, "OK")
-	})
-	r.Head("/-/ready", func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusOK)
-	})
-
-	debugHandlerFunc := func(w http.ResponseWriter, req *http.Request) {
-		subpath := route.Param(req.Context(), "subpath")
-		req.URL.Path = path.Join("/debug", subpath)
-		// path.Join removes trailing slashes, but some pprof handlers expect them.
-		if strings.HasSuffix(subpath, "/") && !strings.HasSuffix(req.URL.Path, "/") {
-			req.URL.Path += "/"
-		}
-		http.DefaultServeMux.ServeHTTP(w, req)
-	}
-	r.Get("/debug/*subpath", debugHandlerFunc)
-	r.Post("/debug/*subpath", debugHandlerFunc)
 }
 
 func disableCaching(w http.ResponseWriter) {
