+++ ui/web_test.go
@@ -14,56 +14,17 @@
 package ui
 
 import (
-	"log/slog"
 	"net/http"
 	"net/http/httptest"
-	"os"
 	"testing"
 
 	"github.com/prometheus/common/route"
 	"github.com/stretchr/testify/require"
 )
 
-func TestDebugHandlersWithRoutePrefix(t *testing.T) {
-	logger := slog.New(slog.NewTextHandler(os.Stderr, nil))
-	reloadCh := make(chan chan error)
-
-	// Test with route prefix
-	routePrefix := "/prometheus/alertmanager"
-	router := route.New().WithPrefix(routePrefix)
-	Register(router, reloadCh, logger)
-
-	// Test GET request to pprof index (note: pprof index returns text/html)
-	req := httptest.NewRequest("GET", routePrefix+"/debug/pprof/", nil)
-	w := httptest.NewRecorder()
-	router.ServeHTTP(w, req)
-
-	require.Equal(t, http.StatusOK, w.Code)
-	require.Contains(t, w.Body.String(), "/debug/pprof/", "pprof page did not load with expected content when using a route prefix")
-
-	// Test GET request to pprof heap endpoint
-	req = httptest.NewRequest("GET", routePrefix+"/debug/pprof/heap", nil)
-	w = httptest.NewRecorder()
-	router.ServeHTTP(w, req)
-
-	require.Equal(t, http.StatusOK, w.Code)
-
-	// Test without route prefix (should also work)
-	router2 := route.New()
-	Register(router2, reloadCh, logger)
-
-	req = httptest.NewRequest("GET", "/debug/pprof/", nil)
-	w = httptest.NewRecorder()
-	router2.ServeHTTP(w, req)
-
-	require.Equal(t, http.StatusOK, w.Code)
-	require.Contains(t, w.Body.String(), "/debug/pprof/", "pprof page did not load with expected content")
-}
-
 func TestWebRoutes(t *testing.T) {
 	router := route.New()
-	logger := slog.New(slog.NewTextHandler(os.Stderr, nil))
-	Register(router, make(chan chan error), logger)
+	Register(router)
 
 	tests := []struct {
 		name         string
